The menu bar is now completed. Every action in the user interface is fully accessible, whether from the main menu bar or from the context menu. Besides, the Edit Menu gets its "delete" menu item updated depending on the current selection: user, account or operation. The accelerator is working, too.
Support for undo/redo actions is being added. It is possible to add, edit and remove accounts and users, undoing and redoing those changes. The undo/redo stack is currently unlimited, but a limit should be established according to the memory available. Adding, editing and removing account operations cannot be undone yet, but work is being done on the matter. 

The actions must return an execution result, which is still missing. Common operations for those actions must be checked in order to fix error handling, and many windows and listeners must be checked, too, to clean up code.